Add a new pic file in directory
When I was previously attempting to project the shrimp pasta picture on the pasta.html page O would only get the alt message. I realized that the image I was attempting to get was not in the same directory as the HTMl page I was attempting to source it from. This prevents that image from being a relative link. So I added a pic file in the recipes directory and added the shrimp picture there